Relocate catalyst's debug module
.. into a top-level sub-package. This ensures it can still be imported
as `import catalyst.debug` or `from catalyst.debug import ...`.
