+++ frontend/catalyst/api_extensions/__init__.py
@@ -16,7 +16,6 @@
 This module is a collection of public API extensions for pragramming with Catalyst frontends.
 """
 
-from catalyst.api_extensions import debug
 from catalyst.api_extensions.callbacks import pure_callback
 from catalyst.api_extensions.control_flow import cond, for_loop, while_loop
 from catalyst.api_extensions.differentiation import (
@@ -31,7 +30,6 @@ from catalyst.api_extensions.function_maps import vmap
 from catalyst.api_extensions.quantum_operators import adjoint, ctrl, measure
 
 __all__ = (
-    "debug",
     "pure_callback",
     "cond",
     "for_loop",
