+++ frontend/catalyst/debug/compiler_functions.py
@@ -12,68 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Catalyst's debug module contains functions useful for user program debugging."""
+"""
+This module contains debug functions to interact with the compiler and compiled functions.
+"""
 
-import builtins
 import os
 
-import jax
 from jax.interpreters import mlir
 
 import catalyst
-from catalyst.api_extensions.callbacks import debug_callback as callback
 from catalyst.compiled_functions import CompiledFunction
 from catalyst.compiler import Compiler
-from catalyst.jax_primitives import print_p
 from catalyst.tracing.contexts import EvaluationContext
 from catalyst.tracing.type_signatures import filter_static_args, promote_arguments
 from catalyst.utils.filesystem import WorkspaceManager
-
-
-# pylint: disable=redefined-builtin
-def print(x, memref=False):
-    """A :func:`qjit` compatible print function for printing values at runtime.
-
-    Enables printing of numeric values at runtime. Can also print objects or strings as constants.
-
-    Args:
-        x (jax.Array, Any): A single jax array whose numeric values are printed at runtime, or any
-            object whose string representation will be treated as a constant and printed at runtime.
-        memref (Optional[bool]): When set to ``True``, additional information about how the array is
-            stored in memory is printed, via the so-called "memref" descriptor. This includes the
-            base memory address of the data buffer, as well as the rank of the array, the size of
-            each dimension, and the strides between elements.
-
-    **Example**
-
-    .. code-block:: python
-
-        @qjit
-        def func(x: float):
-            debug.print(x, memref=True)
-            debug.print("exit")
-
-    >>> func(jnp.array(0.43))
-    Unranked Memref base@ = 0x5629ff2b6680 rank = 0 offset = 0 sizes = [] strides = [] data =
-    [0.43]
-    exit
-
-    Outside a :func:`qjit` compiled function the operation falls back to the Python print statement.
-
-    .. note::
-
-        Python f-strings will not work as expected since they will be treated as Python objects.
-        This means that array values embeded in them will have their compile-time representation
-        printed, instead of actual data.
-    """
-    if EvaluationContext.is_tracing():
-        if isinstance(x, jax.core.Tracer):
-            print_p.bind(x, memref=memref)
-        else:
-            print_p.bind(string=str(x))
-    else:
-        # Dispatch to Python print outside a qjit context.
-        builtins.print(x)
 
 
 def print_compilation_stage(fn, stage):
@@ -185,12 +137,3 @@ def compile_from_mlir(ir, compiler=None, compile_options=None):
         result_types = [mlir.ir.RankedTensorType.parse(rt) for rt in func_data[1].split(",")]
 
     return CompiledFunction(shared_object, qfunc_name, result_types, compiler.options)
-
-
-__all__ = (
-    "callback",
-    "print",
-    "print_compilation_stage",
-    "get_cmain",
-    "compile_from_mlir",
-)
